Create you-guys-just.md
I'm loath to include race or gender in descriptions unless they contribute to the meme themself. People should be just people first and foremost, and characteristics such as that should not factor in unless the issue is relevant to the image -- and especially, when it comes to gender, we cannot assume.